pyproject.toml
[tool.poetry]
name = "fastapi-uploadfiles"
version = "0.1.0"
description = ""
authors = ["leonidig <leonid200955@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = "^0.32.1"
python-multipart = "^0.0.18"
pillow = "^11.0.0"
httpx = "^0.28.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
